+++ src/enum.js
@@ -2,8 +2,19 @@ const Filter = {
     ALL: 'all',
     POPULAR: 'popular',
     NEW: 'new'
-}
+};
+
+const Locale = {
+    English: 'en',
+    Russian: 'ru'
+};
+
+const Extention = {
+    Jpg: '.jpg'
+};
 
 export {
-    Filter
-}
+    Filter,
+    Locale,
+    Extention
+};