+++ src/filters.js
@@ -4,15 +4,15 @@ import { renderArticleList } from './article-list.js';
 const sortListElement = document.querySelector('.sort-list');
 const articleListElement = document.getElementById('app');
 
-const compareDatesArticles = (articleA, articleB) => {
-    const dateA = new Date(articleA.created_at);
-    const dateB = new Date(articleB.created_at);
+// деструктуризация объекта
+const compareArticlesByDate = ({created_at: createdA}, {created_at: createdB}) => {
+    const dateA = new Date(createdA);
+    const dateB = new Date(createdB);
 
     return dateB.getTime() - dateA.getTime();
 }
 
-const comparePopularityArticels = (articleA, articleB) => {
-
+const compareArticlesByPopularity = (articleA, articleB) => {
     return (articleB.likes + articleB.comments) - (articleA.likes + articleA.comments)
 }
 
@@ -33,10 +33,10 @@ const getFilterClickHandler = (articles) => {
                 renderArticleList(articles, articleListElement);
                 break;
             case Filter.POPULAR:
-                renderArticleList(articles.slice().sort(comparePopularityArticels), articleListElement);
+                renderArticleList(articles.slice().sort(compareArticlesByPopularity), articleListElement);
                 break;
             case Filter.NEW:
-                renderArticleList(articles.slice().sort(compareDatesArticles), articleListElement);
+                renderArticleList(articles.slice().sort(compareArticlesByDate), articleListElement);
                 break;
         }
     };